Enum mpl_auction::errors::AuctionError

pub enum AuctionError {

    IncorrectOwner,
    NotRentExempt,
    InvalidBidAccount,
    InvalidAuctionAccount,
    BalanceTooLow,
    InvalidState,
    BidTooSmall,
    AuctionTransitionInvalid,
    DerivedKeyInvalid,
    TokenTransferFailed,
    TokenMintToFailed,
    TokenBurnFailed,
    InvalidAuthority,
    AuthorityNotSigner,
    NumericalOverflowError,
    BidderPotTokenAccountOwnerMismatch,
    Uninitialized,
    MetadataInvalid,
    BidderPotDoesNotExist,
    BidAlreadyActive,
    IncorrectMint,
    MustReveal,
    InvalidReveal,
    BidderPotEmpty,
    InvalidTokenProgram,
    DelegateShouldBeNone,
    CloseAuthorityShouldBeNone,
    DataTypeMismatch,
    BidMustBeMultipleOfTickSize,
    GapBetweenBidsTooSmall,
    InvalidGapTickSizePercentage,
    BidderPotTokenAccountMustBeNew,

}

Errors that may be returned by the Auction program.

Variants

IncorrectOwner

Account does not have correct owner

NotRentExempt

Lamport balance below rent-exempt threshold.

InvalidBidAccount

Bid account provided does not match the derived address.

InvalidAuctionAccount

Auction account specified is invalid.

BalanceTooLow

Balance too low to make bid.

InvalidState

Auction is not currently running.

BidTooSmall

Bid is too small.

AuctionTransitionInvalid

Invalid transition, auction state may only transition: Created -> Started -> Stopped

DerivedKeyInvalid

Failed to derive an account from seeds.

TokenTransferFailed

Token transfer failed

TokenMintToFailed

Token mint to failed

TokenBurnFailed

Token burn failed

InvalidAuthority

Invalid authority

AuthorityNotSigner

Authority not signer

NumericalOverflowError

Numerical overflow

BidderPotTokenAccountOwnerMismatch

Bidder pot token account does not match

Uninitialized

Uninitialized

MetadataInvalid

Metadata account is missing or invalid.

BidderPotDoesNotExist

Bidder pot is missing, and required for SPL trades.

BidAlreadyActive

Existing Bid is already active.

IncorrectMint

Incorrect mint specified, must match auction.

MustReveal

Must reveal price when ending a blinded auction.

InvalidReveal

The revealing hash is invalid.

BidderPotEmpty

The pot for this bid is already empty.

InvalidTokenProgram

This is not a valid token program

DelegateShouldBeNone

Accept payment delegate should be none

CloseAuthorityShouldBeNone

Accept payment close authority should be none

DataTypeMismatch

Data type mismatch

BidMustBeMultipleOfTickSize

Bid must be multiple of tick size

GapBetweenBidsTooSmall

During the gap window, gap between next lowest bid must be of a certain percentage

InvalidGapTickSizePercentage

Gap tick size percentage must be between 0 and 100

BidderPotTokenAccountMustBeNew

Gap tick size percentage must be between 0 and 100