Module mpl_auction::processor

Re-exports

pub use cancel_bid::*;

pub use claim_bid::*;

pub use create_auction::*;

pub use create_auction_v2::*;

pub use end_auction::*;

pub use place_bid::*;

pub use set_authority::*;

pub use start_auction::*;

Modules

cancel_bid

Cancels an existing bid. This only works in two cases:

claim_bid

Claim bid winnings into a target SPL account, only the authorised key can do this, though the target can be any SPL account.

create_auction

create_auction_v2

end_auction

place_bid

Places a bid on a running auction, the logic here implements a standard English auction mechanism, once the auction starts, new bids can be made until 10 minutes has passed with no new bid. At this point the auction ends.

set_authority

Resets authority on an auction account.

start_auction

Structs

AuctionData

AuctionDataExtended

Bid

Bids associate a bidding key with an amount bid.

BidderMetadata

Models a set of metadata for a bidder, meant to be stored in a PDA. This allows looking up information about a bidder regardless of if they have won, lost or cancelled.

BidderPot

Enums

AuctionState

Define valid auction state transitions.

BidState

BidState tracks the running state of an auction, each variant represents a different kind of auction being run.

PriceFloor

Structure with pricing floor data.

WinnerLimit

Constants

BASE_AUCTION_DATA_SIZE

BIDDER_METADATA_LEN

BID_LENGTH

MAX_AUCTION_DATA_EXTENDED_SIZE

Functions

process_instruction

Type Definitions

AuctionName